+++ services/appinsights/mgmt/2020-02-02/insights/models.go
@@ -210,7 +210,7 @@ type ApplicationInsightsComponentAnalyticsItem struct {
 	Version *string `json:"Version,omitempty"`
 	// Scope - Enum indicating if this item definition is owned by a specific user or is shared between all users with access to the Application Insights component. Possible values include: 'ItemScopeShared', 'ItemScopeUser'
 	Scope ItemScope `json:"Scope,omitempty"`
-	// Type - Enum indicating the type of the Analytics item. Possible values include: 'ItemTypeQuery', 'ItemTypeFunction', 'ItemTypeFolder', 'ItemTypeRecent'
+	// Type - Enum indicating the type of the Analytics item. Possible values include: 'ItemTypeNone', 'ItemTypeQuery', 'ItemTypeRecent', 'ItemTypeFunction'
 	Type ItemType `json:"Type,omitempty"`
 	// TimeCreated - READ-ONLY; Date and time in UTC when this item was created.
 	TimeCreated *string `json:"TimeCreated,omitempty"`