+++ services/preview/appinsights/mgmt/2021-11-01-preview/insights/models.go
@@ -197,7 +197,7 @@ type ApplicationInsightsComponentAnalyticsItem struct {
 	Version *string `json:"Version,omitempty"`
 	// Scope - Enum indicating if this item definition is owned by a specific user or is shared between all users with access to the Application Insights component. Possible values include: 'ItemScopeShared', 'ItemScopeUser'
 	Scope ItemScope `json:"Scope,omitempty"`
-	// Type - Enum indicating the type of the Analytics item. Possible values include: 'ItemTypeQuery', 'ItemTypeFunction', 'ItemTypeFolder', 'ItemTypeRecent'
+	// Type - Enum indicating the type of the Analytics item. Possible values include: 'ItemTypeNone', 'ItemTypeQuery', 'ItemTypeRecent', 'ItemTypeFunction'
 	Type ItemType `json:"Type,omitempty"`
 	// TimeCreated - READ-ONLY; Date and time in UTC when this item was created.
 	TimeCreated *string `json:"TimeCreated,omitempty"`
@@ -2528,7 +2528,7 @@ type Workbook struct {
 	SystemData *SystemData `json:"systemData,omitempty"`
 	// Identity - Identity used for BYOS
 	Identity *WorkbookResourceIdentity `json:"identity,omitempty"`
-	// Kind - The kind of workbook. Choices are user and shared. Possible values include: 'KindUser', 'KindShared'
+	// Kind - The kind of workbook. Only valid value is shared. Possible values include: 'KindUser', 'KindShared'
 	Kind Kind `json:"kind,omitempty"`
 	// Etag - Resource etag
 	Etag map[string]*string `json:"etag"`
@@ -2783,7 +2783,7 @@ type WorkbookPropertiesUpdateParameters struct {
 type WorkbookResource struct {
 	// Identity - Identity used for BYOS
 	Identity *WorkbookResourceIdentity `json:"identity,omitempty"`
-	// Kind - The kind of workbook. Choices are user and shared. Possible values include: 'KindUser', 'KindShared'
+	// Kind - The kind of workbook. Only valid value is shared. Possible values include: 'KindUser', 'KindShared'
 	Kind Kind `json:"kind,omitempty"`
 	// Etag - Resource etag
 	Etag map[string]*string `json:"etag"`
@@ -3283,7 +3283,7 @@ func (wtup *WorkbookTemplateUpdateParameters) UnmarshalJSON(body []byte) error {
 // WorkbookUpdateParameters the parameters that can be provided when updating workbook properties
 // properties.
 type WorkbookUpdateParameters struct {
-	// Kind - The kind of workbook. Choices are user and shared. Possible values include: 'SharedTypeKindUser', 'SharedTypeKindShared'
+	// Kind - The kind of workbook. Only valid value is shared. Possible values include: 'SharedTypeKindUser', 'SharedTypeKindShared'
 	Kind SharedTypeKind `json:"kind,omitempty"`
 	// Tags - Resource tags.
 	Tags map[string]*string `json:"tags"`