+++ services/appinsights/mgmt/2015-05-01/insights/CHANGELOG.md
@@ -1,2 +1,13 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Removed Constants
+
+1. ItemType.Folder
+
+## Additive Changes
+
+### New Constants
+
+1. ItemType.None