+++ services/appinsights/mgmt/2020-02-02/insights/CHANGELOG.md
@@ -1,2 +1,13 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Removed Constants
+
+1. ItemType.ItemTypeFolder
+
+## Additive Changes
+
+### New Constants
+
+1. ItemType.ItemTypeNone