+++ services/preview/appinsights/mgmt/2022-01-11-preview/insights/CHANGELOG.md
@@ -1,2 +1,16 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Funcs
+
+1. WorkbooksClient.Get
+	- Params
+		- From: context.Context, string, string
+		- To: context.Context, string, string, *bool
+1. WorkbooksClient.GetPreparer
+	- Params
+		- From: context.Context, string, string
+		- To: context.Context, string, string, *bool